Refactor IntroModal to use useImagePreloader
